docs: Added Access Control
Added Access Control to the Job Controller limiting the actions a user can do to create, edit, delete. User must be logged in as themselves in order to create, edit, delete.

Adding Access Control to the Job Controller makes it so any user can not just edit or delete a specific users posting.

+++ controllers/JobController.php
@@ -11,10 +11,28 @@ use app\models\Category;
 use app\models\Job;
 
 class JobController extends \yii\web\Controller{
+    /*
+    * Access Control
+    */
+    public function behaviors(){
+        return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['create', 'edit', 'delete'],
+                'rules' => [
+                    [
+                        'actions' => ['create', 'edit', 'delete'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ]
+        ];
+    }
+
     public function actionIndex(){
         //   Create Query
         $query = Job::find();
-
         $pagination = new Pagination([
             'defaultPageSize' => 20,
             'totalCount' => $query->count(),
@@ -24,7 +42,6 @@ class JobController extends \yii\web\Controller{
             ->offset($pagination->offset)
             ->limit($pagination->limit)
             ->all();
-
         return $this->render('index', [
             'jobs' => $jobs,
             'pagination' => $pagination,
@@ -36,27 +53,22 @@ class JobController extends \yii\web\Controller{
         $job = Job::find()
             ->where(['id' => $id])
             ->one();
-
         //  Render View
         return $this->render('details', ['job' => $job]);
     }
 
     public function actionCreate(){
         $job = new Job();
-
        if ($job->load(Yii::$app->request->post())) {
            if ($job->validate()) {
             // Save
             $job->save();
-
             // Show Message
             Yii::$app->getSession()->setFlash('success', 'Job Added');
-
             // Redirect
             return $this->redirect('index.php?r=job');
            }
        }
-
        return $this->render('create', [
            'job' => $job,
        ]);
@@ -64,32 +76,36 @@ class JobController extends \yii\web\Controller{
 
     public function actionDelete($id){
         $job = Job::findOne($id);
-
+        // check for ownership
+        if(Yii::$app->user->identity->id != $job->user_id){
+            // Redirect
+            return $this->redirect('index.php?r=job');
+        }
+        // Delete Job
         $job->delete();
-
         // Show Message
         Yii::$app->getSession()->setFlash('success', 'Job Deleted');
-
         // Redirect
         return $this->redirect('index.php?r=job');
     }
 
     public function actionEdit($id){
         $job = Job::findOne($id);
-
+        // check for ownership
+        if(Yii::$app->user->identity->id != $job->user_id){
+            // Redirect
+            return $this->redirect('index.php?r=job');
+        }
        if ($job->load(Yii::$app->request->post())) {
            if ($job->validate()) {
             // Save
             $job->save();
-
             // Show Message
             Yii::$app->getSession()->setFlash('success', 'Job Updated');
-
             // Redirect
             return $this->redirect('index.php?r=job');
            }
        }
-
        return $this->render('edit', [
            'job' => $job,
        ]);
